requirements2.txt
numpy
Pillow
matplotlib
